Item1626: not quite perfect simulation of the rendering environment meant META{formfield was not expanded correctly
git-svn-id: http://svn.foswiki.org/trunk/PublishPlugin@3920 0b4bb1d4-4e5a-0410-9cc4-b2b747904278

+++ lib/Foswiki/Plugins/PublishPlugin/file.pm
@@ -31,6 +31,7 @@ sub new {
 
     foreach my $param qw(defaultpage googlefile relativeurl) {
         my $p = $query->param($param);
+        die $param unless defined $p;
         $p =~ /^(.*)$/;
         $this->{$param} = $1;
         $query->delete($param);
